pyproject.toml
[tool.poetry]
name = "cookbook"
version = "0.1.0"
description = ""
authors = ["Lana <lan2828@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
